requirements.txt
python-telegram-bot==20.7
openai-whisper==20231117
qdrant-client==1.7.0
sentence-transformers==2.2.2
openai==1.12.0
python-dotenv==1.0.0
torch==2.1.2
numpy<2.0.0
